[1/3] Use AndroidJUnit4 instead of JUnit4
The JUnit4 test runner does not respect the timeout argument given to
the instrumentation via timeout_msec, which we use on our build server
to enforce maximum duration of small and medium tests.

The AndroidJUnit4 runner does use this argument.

Suppress lint for existing uses of unsupported test runners that can't
be converted to AndroidJUnit4.

Bug: 166771118
Test: ./gradlew bOS
Change-Id: I8c0d1eb66a78834d301f7a81c3f9baa5328af2fa