Add a defaultValue val to the Serializer interface to make implementing Serializers easier and clearer and add a check to make sure that the user does not close the output stream from writeTo.
Test: Added a test for default value and all existing tests pass
Bug: 168027417
Relnote: "Serializers now require a new property for the default value that will be used if there is no data on disk. This makes it easier to implement custom serializers so users do not have to special case empty input streams (empty input streams don't parse with json).

Also there is now a check to confirm that the output stream provided to writeTo() is not closed, and if it is closed it throws exceptions"

Change-Id: I16e2984b16cb387d8df55bad7a82a8120e94597a